requirements.txt
fastapi
uvicorn
python-dotenv
tripo3d
